requirements.txt
requests
PyCrypto
discord.py[voice]
